requirements.txt
wikipedia
python-telegram-bot~=20.7
requests